разбивает проект на модули и создает функцию проверки времени встречи
