+++ index.html
@@ -235,5 +235,6 @@
     </template>
 
 </body>
-<script type="text/javascript" src="js/main.js"></script>
+<script type="module" src="js/main.js"></script>
+<script type="text/javascript" src="js/functions.js"></script>
 </html>
