+++ js/functions.js
@@ -1,4 +1,4 @@
-
+const msInMin = 60000;
 const checksLengthString = (checkString, maxLength) => checkString.length <= maxLength;
 
 checksLengthString('Пропоганда', 15);
@@ -24,3 +24,18 @@ const getNumberFromString = (checkString) => {
 };
 
 getNumberFromString('524 проблемы и 38 решений');
+
+const getTimeInMinutes = (time) => (new Date(0, 0, 0, ...time.split(':')) / msInMin);
+
+const isMeetTimeAppropriate = (startWorkingTime, endWorkinTime, startMeetingTime, meetingDuration) => {
+
+  const startWork = getTimeInMinutes(startWorkingTime);
+  const endWork = getTimeInMinutes(endWorkinTime);
+  const startMeeting = getTimeInMinutes(startMeetingTime);
+  if (startMeeting < startWork) {
+    return false;
+  }
+  return (endWork - startMeeting) >= meetingDuration;
+};
+
+isMeetTimeAppropriate('8:00', '17:30', '13:00', 1000);
